split apache arrow flight and dremio plugins
