+++ plugin-jdbc-arrow-flight/build.gradle
@@ -1,11 +1,11 @@
-project.description = 'Query Dremio databases using the Kestra JDBC plugin.'
+project.description = 'Query Dremio databases using the Kestra Apache Arrow Flight JDBC plugin.'
 
 jar {
     manifest {
         attributes(
                 "X-Kestra-Name": project.name,
-                "X-Kestra-Title": "Dremio",
-                "X-Kestra-Group": project.group + ".jdbc.dremio",
+                "X-Kestra-Title": "Arrow Flight",
+                "X-Kestra-Group": project.group + ".jdbc.arrowflight",
                 "X-Kestra-Description": project.description,
                 "X-Kestra-Version": project.version
         )
@@ -13,7 +13,7 @@ jar {
 }
 
 dependencies {
-    implementation("com.dremio.distribution:dremio-jdbc-driver:3.0.6-201812082352540436-1f684f9")
+    implementation("org.apache.arrow:flight-sql-jdbc-driver:12.0.1")
     implementation project(':plugin-jdbc')
 
     testImplementation project(':plugin-jdbc').sourceSets.test.output
