+++ plugin-jdbc-arrow-flight/src/main/java/io/kestra/plugin/jdbc/arrowflight/package-info.java
@@ -1,0 +1,7 @@
+@PluginSubGroup(
+    description = "This sub-group of plugins contains tasks for accessing the databases through Apache Arrow Flight.",
+    categories = PluginSubGroup.PluginCategory.DATABASE
+)
+package io.kestra.plugin.jdbc.arrowflight;
+
+import io.kestra.core.models.annotations.PluginSubGroup;
